requirements.txt
telethon
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2